requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.1.4
langchain-google-genai==0.0.6
langchain-community==0.0.16